fixed cleaning script
clean.py now removes files correctly and doesnt remove .venv

+++ clean.py
@@ -1,4 +1,5 @@
 import os # preinstalled
 
-for file in ["data.json", "embeddings.db", ".venv", "config.cfg"]:
-    os.remove(file)
+for file in ["data.json", "embeddings.db", "config.cfg"]:
+    if os.path.exists(file):
+        os.remove(file)
